Override CVX groups for certain CVX codes non in CDC group mapping
The impetus for this change is a DFS student with four 182 OPV shots that is
showing up non-compliant. When parsing our Maine imm record, we group them
under Polio because Maine tells us to. However, in the ICE transformations, we
map substance code 182 to group code 182 because that's the group that appears
in our mapping file. We then fail to join the events on group code and mark the
dose validity for these doses as seen by ICE as null/"Unknown".

The problem is that the CDC group mapping does not include all CVX codes in its
group mapping file we download in the Rake task. For those vaccine codes, we
simply create self-referential CVX groups under the "groups" key.

I manually compiled the full list of these CVX codes and have modified the Rake
task to override the groups for these codes where it makes sense:

|Code|Description|Thoughts|
|-|-|-|
|169|Hep A, live attenuated|We should pin this to the Hep A group|
|167|meningococcal, unknown serogroups|I don't think we can touch this one since we don't know if it should go to MenACWY or MenB|
|78|tularemia vaccine|NA|
|68|melanoma|NA|
|61|HIV never active|NA|
|59|Hep E|NA|
|11|pertussis|Can't pin this to 107 or 115 since it is missing tetanus|
|179|OPV ,monovalent, unspecified|We should pin this to Polio|
|67|malaria|NA|
|58|Hep C|NA|
|63|Junin virus|NA|
|131|typhus, historical|NA|
|192|meningococcal AC polysaccharide (non-US)|Pin to MenACWY|
|191|meningococcal A polysaccharide (non-US)|Pin to MenACWY|
|190|Typhoid conjugate vaccine (TCV)|NA|
|57|hantavirus|NA|
|64|leishmaniasis|NA|
|69|parainfluenza-3|NA|
|178|OPV bivalent|Pin to polio|
|182|OPV, Unspecified|Pin to polio|
|142|tetanus toxoid, not adsorbed|Pin to Td|
|72|rheumatic fever|NA|
|112|tetanus toxoid, unspecified formulation|Pin to Td|
|198|DTP-hepB-Hib Pentavalent Non-US|Pin to DTAP _and_ HepB _and_ Hib|
|60|herpes simplex 2|NA|
|70|Q Fever|NA|
|65|leprosy|NA|
|73|Rift Valley fever|NA|